Display confliction revisions
- list view in ListConflicts.java (show the title and updated_at properties), the user can pick the winning revision
- clean up unused methods of Facebook SDK in MainActivity.java
- move TaskFragment to its own file

https://github.com/couchbaselabs/ToDoLite-Android/issues/35
