+++ ToDoLite/src/main/AndroidManifest.xml
@@ -34,6 +34,9 @@
         <activity
             android:name=".LoginActivity">
         </activity>
+        <activity
+            android:name=".ListConflicts">
+        </activity>
     </application>
 
 </manifest>
