+++ ToDoLite/src/main/java/com/couchbase/todolite/LoginActivity.java
@@ -1,13 +1,10 @@
 package com.couchbase.todolite;
 
-import android.app.Activity;
 import android.app.AlertDialog;
-import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v4.app.FragmentActivity;
-import android.util.AttributeSet;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
