+++ ToDoLite/src/main/res/layout/view_task.xml
@@ -30,6 +30,12 @@
             android:layout_gravity="center_vertical"
             android:singleLine="false"
             android:textColor="@color/primary_text_color"/>
+        <ImageView
+            android:id="@+id/conflict_icon"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:src="@drawable/ic_call_split_black_24dp"
+            android:visibility="invisible"/>
         <CheckBox
             android:id="@+id/checked"
             android:layout_width="wrap_content"
