+++ ToDoLite/src/main/res/values/strings.xml
@@ -8,11 +8,13 @@
     <string name="action_share">Share</string>
     <string name="action_delete">Delete</string>
     <string name="action_show_document">Show Document</string>
+    <string name="action_update">Update</string>
     <string name="hint_new_list">List title</string>
     <string name="hint_create_task">Add new task</string>
     <string name="navigation_drawer_open">Open drawer</string>
     <string name="navigation_drawer_close">Close drawer</string>
     <string name="title_activity_share">Share</string>
     <string name="title_dialog_new_list">Create new list</string>
+    <string name="title_dialog_update">Update task</string>
     <string name="title_activity_image_view">ImageViewActivity</string>
 </resources>
